+++ src/BVH/BVH.h
@@ -26,7 +26,7 @@ namespace Raycer
 		void build(std::vector<Triangle>& triangles, std::vector<TriangleSOA<4>>& triangles4);
 		bool intersect(const Scene& scene, const Ray& ray, Intersection& intersection) const;
 
-		static BVHSplitOutput calculateSplit(std::vector<BVHBuildTriangle>& buildTriangles, std::vector<BVHSplitCache>& cache, uint64_t start, uint64_t end);
+		static BVHSplitOutput calculateSplit(std::vector<BVHBuildTriangle>& buildTriangles, std::vector<BVHSplitCache>& cache, uint32_t start, uint32_t end);
 
 		BVHType type = BVHType::BVH1;
